Detect and display features available + elapsed time

+++ example/www/openscad-runner.js
@@ -1,24 +1,29 @@
-
+// Output is {outputs: [name, content][], mergedOutputs: [{(stderr|stdout|error)?: string}], exitCode: number}
 export function spawnOpenSCAD({inputs, args, outputPaths}) {
   var worker;
   var rejection;
+
+  function terminate() {
+    if (!worker) {
+      return;
+    }
+    worker.terminate();
+    worker = null;
+  }
     
   const promise = new Promise((resolve, reject) => {
     worker = new Worker('./openscad-worker.js', {'type': 'module'});
     rejection = reject;
     worker.onmessage = e => {
-      // if (e.data.error) {
-      //   reject(e.data);
-      // } else {
-        resolve(e.data);
-      // }
-    };
+      resolve(e.data);
+      terminate();
+    }
     worker.postMessage({inputs, args, outputPaths})
   });
   
   promise.kill = () => {
-    worker.terminate();
-    rejection({error: 'Killed'});
+    rejection({error: 'Terminated'});
+    terminate();
   }
 
   return promise;
